% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIAdb_connect.R
\name{FIAdb_connect}
\alias{FIAdb_connect}
\title{Connect to an FIA PostgreSQL Database}
\usage{
FIAdb_connect(dbname = "fiadb")
}
\arguments{
\item{dbname}{PostgreSQL database name.}
}
\value{
A DBI database connection object.
}
\description{
Creates a connection to an FIA PostgreSQL database using the
supplied database name and user credentials.
}
\keyword{internal}
