% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuery.R
\name{getQuery}
\alias{getQuery}
\title{Execute an FIA Database Query}
\usage{
getQuery(query, con)
}
\arguments{
\item{query}{Character string containing a valid SQL query.}

\item{con}{A DBI database connection object.}
}
\value{
A data frame containing the results of the SQL query.
}
\description{
Executes an SQL query using an existing database connection and
returns the results as a data frame.
}
\details{
This function is primarily intended for internal package use and is
used by other functions when retrieving FIA database records and
metadata.
}
\examples{
\dontrun{
con <- FIAdb_connect()

getQuery(
  "SELECT * FROM fs_fiadb.plot LIMIT 10",
  con
)
}

}
\keyword{internal}
